Refactor: Extract fetchAnalyticsData utility and update analytics integration
