+++ site/src/components/TurtlesBackground.tsx
@@ -1,21 +1,23 @@
 import AnalyticData from '@/components/AnalyticData'
 
 export default function TurtlesBackground({
-  header,
-  initialVolume,
-}: {
+                                            header,
+                                            initialVolume,
+                                          }: {
   header: string
   initialVolume: number | undefined
 }) {
   return (
     <div className="absolute top-0 z-0">
-      <div className="turtle-background flex h-[78vh] w-screen flex-col items-center justify-center overflow-hidden bg-[url('/bg.png')] bg-cover bg-bottom">
+      <div
+        className="turtle-background flex h-[78vh] w-screen flex-col items-center justify-center overflow-hidden bg-[url('/bg.png')] bg-cover bg-bottom">
         <div className="turtle-dark-overlay flex w-screen flex-col items-center justify-center">
           <div className="m-[4vw] flex-col items-center justify-center">
-            <h1 className="turtle-text-shadow m-0 text-center text-[11vw] leading-[100%] text-white sm:text-[9vw] 3xl:text-[10rem]">
+            <h1
+              className="turtle-text-shadow m-0 text-center text-[11vw] leading-[100%] text-white sm:text-[9vw] 3xl:text-[10rem]">
               {header}
             </h1>
-            {initialVolume && <AnalyticData initialVolume={initialVolume} />}
+            <AnalyticData initialVolume={initialVolume} />
           </div>
         </div>
       </div>
